XRDDEV-2082 Break up generic "/clients" API
