+++ src/centralserver/admin-service/src/test/java/org/niis/xroad/centralserver/restapi/openapi/ClientsApiTest.java
@@ -4,17 +4,17 @@
  * Copyright (c) 2018 Estonian Information System Authority (RIA),
  * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
  * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -25,41 +25,28 @@
  */
 package org.niis.xroad.centralserver.restapi.openapi;
 
-import org.apache.commons.lang3.StringUtils;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInfo;
 import org.niis.xroad.centralserver.openapi.model.ClientDto;
-import org.niis.xroad.centralserver.openapi.model.ClientIdDto;
-import org.niis.xroad.centralserver.openapi.model.MemberNameDto;
 import org.niis.xroad.centralserver.openapi.model.PagedClientsDto;
-import org.niis.xroad.centralserver.openapi.model.XRoadIdDto;
 import org.niis.xroad.centralserver.restapi.repository.FlattenedSecurityServerClientViewRepositoryTest;
 import org.niis.xroad.centralserver.restapi.util.TestUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.test.context.support.WithMockUser;
 
-import java.time.OffsetDateTime;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.springframework.http.HttpStatus.NOT_FOUND;
 
-class ClientsApiTest extends AbstractApiRestTemplateTestContext {
+public class ClientsApiTest extends AbstractApiRestTemplateTestContext {
 
     private static final String PATH = "/api/v1/clients";
-    public static final String MEMBER_M1_CLIENT_ID = "TEST:GOV:M1";
-    public static final String MEMBER_M1_NAME = "Member1";
 
     @Autowired
     TestRestTemplate restTemplate;
@@ -67,34 +54,6 @@ class ClientsApiTest extends AbstractApiRestTemplateTestContext {
     @BeforeEach
     void addApiKeyAuthorizationHeader() {
         TestUtils.addApiKeyAuthorizationHeader(restTemplate);
-    }
-
-    @Nested
-    @DisplayName("POST " + PATH)
-    class AddClient {
-
-        @Test
-        @WithMockUser(authorities = {"ADD_NEW_MEMBER"})
-        @DisplayName("should add client")
-        public void addClient(TestInfo testInfo) {
-            String memberCode = composeMemberName(testInfo);
-            ClientIdDto memberIdDto =  (ClientIdDto) new ClientIdDto()
-                    .memberClass("GOV")
-                    .memberCode(memberCode)
-                    .instanceId("TEST")
-                    .type(XRoadIdDto.TypeEnum.MEMBER);
-            ClientDto memberDto = new ClientDto()
-                    .memberName("memberName")
-                    .xroadId(memberIdDto);
-
-            ResponseEntity<ClientDto> response = restTemplate.postForEntity(PATH, memberDto, ClientDto.class);
-
-            assertNotNull(response);
-            assertEquals(HttpStatus.OK, response.getStatusCode());
-            assertNotNull(response.getBody());
-            assertEquals("memberName", response.getBody().getMemberName());
-            assertEquals(memberIdDto, response.getBody().getXroadId()); // fixme: XRDDEV-
-        }
     }
 
     @Nested
@@ -107,7 +66,7 @@ class ClientsApiTest extends AbstractApiRestTemplateTestContext {
             uriVariables.put("securityServerId", "TEST:GOV:M1:server1");
 
             ResponseEntity<PagedClientsDto> response = restTemplate.getForEntity(
-                    "/api/v1/clients?security_server={securityServerId}",
+                    PATH + "?security_server={securityServerId}",
                     PagedClientsDto.class,
                     uriVariables);
 
@@ -120,7 +79,7 @@ class ClientsApiTest extends AbstractApiRestTemplateTestContext {
         @Test
         void searchAll() {
             ResponseEntity<PagedClientsDto> response = restTemplate.getForEntity(
-                    "/api/v1/clients", PagedClientsDto.class);
+                    PATH, PagedClientsDto.class);
 
             assertNotNull(response);
             assertEquals(200, response.getStatusCodeValue());
@@ -135,19 +94,19 @@ class ClientsApiTest extends AbstractApiRestTemplateTestContext {
             uriVariables.put("sort", "xroad_id.instance_id");
             uriVariables.put("desc", "false");
             ResponseEntity<PagedClientsDto> response = restTemplate.getForEntity(
-                    "/api/v1/clients?sort={sort}&desc={desc}", PagedClientsDto.class, uriVariables);
+                    PATH + "?sort={sort}&desc={desc}", PagedClientsDto.class, uriVariables);
             assertEquals("Instance2", response.getBody().getClients().get(0).getXroadId().getInstanceId());
 
             uriVariables.put("desc", "true");
             response = restTemplate.getForEntity(
-                    "/api/v1/clients?sort={sort}&desc={desc}", PagedClientsDto.class, uriVariables);
+                    PATH + "?sort={sort}&desc={desc}", PagedClientsDto.class, uriVariables);
             assertEquals("TEST", response.getBody().getClients().get(0).getXroadId().getInstanceId());
         }
 
         @Test
         void stableSort() {
             // check that we always have secondary sort by id, to guarantee e.g. stable paging
-            String clientSearchUrl = "/api/v1/clients?sort={sort}&desc={desc}&instance={instance}";
+            String clientSearchUrl = PATH + "?sort={sort}&desc={desc}&instance={instance}";
             var uriVariables = new HashMap<String, String>();
             uriVariables.put("sort", "xroad_id.instance_id");
             uriVariables.put("desc", "false");
@@ -210,101 +169,6 @@ class ClientsApiTest extends AbstractApiRestTemplateTestContext {
                 previousClientId = clientId;
             }
         }
-    }
-
-    @Nested
-    @DisplayName("GET " + PATH + "/{id}")
-    @WithMockUser(authorities = {"VIEW_MEMBER_DETAILS"})
-    class GetClient {
-        private final ClientIdDto expectedMemberId = (ClientIdDto) new ClientIdDto()
-                .memberClass("GOV")
-                .memberCode("M1")
-                .instanceId("TEST")
-                .type(XRoadIdDto.TypeEnum.MEMBER);;
-
-        @Test
-        void getClient() {
-            var uriVariables = Map.of("clientId", MEMBER_M1_CLIENT_ID);
-
-            var response = restTemplate.getForEntity(
-                    "/api/v1/clients/{clientId}",
-                    ClientDto.class,
-                    uriVariables);
-
-            assertNotNull(response);
-            assertEquals(200, response.getStatusCodeValue());
-            assertNotNull(response.getBody());
-            assertEquals(MEMBER_M1_NAME, response.getBody().getMemberName());
-            assertEquals(expectedMemberId, response.getBody().getXroadId());
-        }
-    }
-
-    @Nested
-    @DisplayName("DELETE " + PATH + "/{id}")
-    @WithMockUser(authorities = {"DELETE_MEMBER"})
-    class DeleteClient {
-
-        @Test
-        void deleteClient() {
-            ClientIdDto memberIdDto = (ClientIdDto) new ClientIdDto()
-                    .memberClass("GOV")
-                    .memberCode("to-be-deleted")
-                    .instanceId("TEST")
-                    .type(XRoadIdDto.TypeEnum.MEMBER);
-            ClientDto memberDto = new ClientDto()
-                    .memberName("To Be Deleted")
-                    .xroadId(memberIdDto);
-            restTemplate.postForEntity(PATH, memberDto, ClientDto.class);
-            var uriVariables = Map.of("clientId", "TEST:GOV:to-be-deleted");
-
-            restTemplate.delete(
-                    "/api/v1/clients/{clientId}",
-                    uriVariables);
-
-            ResponseEntity<ClientDto> deletedClientResponse =
-                    restTemplate.getForEntity("/api/v1/clients/{clientId}", ClientDto.class, uriVariables);
-
-            assertNotNull(deletedClientResponse.getBody());
-            assertEquals(NOT_FOUND, deletedClientResponse.getStatusCode());
-        }
-    }
-
-    @Nested
-    @DisplayName("PATCH" + PATH + "/{id}")
-    @WithMockUser(authorities = {"EDIT_MEMBER_NAME_AND_ADMIN_CONTACT"})
-    class UpdateMemberName {
-
-        @Test
-        void updateMemberName() {
-            var timeBeforeUpdate = OffsetDateTime.now();
-
-            ClientDto response = doUpdateName("new name");
-
-            assertNotNull(response);
-            assertEquals("new name", response.getMemberName());
-            assertTrue(response.getUpdatedAt().isAfter(timeBeforeUpdate), "Expecting updatedAt to be updated");
-        }
-
-        private ClientDto doUpdateName(String newName) {
-            var uriVariables = Map.of("clientId", MEMBER_M1_CLIENT_ID);
-
-            return restTemplate.patchForObject(
-                    "/api/v1/clients/{clientId}",
-                    new MemberNameDto().memberName(newName),
-                    ClientDto.class,
-                    uriVariables);
-        }
-
-        @AfterEach
-        void revertNameUpdate() {
-            doUpdateName(MEMBER_M1_NAME);
-        }
-    }
-
-    private static String composeMemberName(TestInfo testInfo) {
-        String testClassSimpleName = StringUtils.substringAfterLast(testInfo.getTestClass().get().getName(), ".");
-        String testMethodName = testInfo.getTestMethod().get().getName();
-        return testClassSimpleName + "." + testMethodName;
     }
 
 }
