+++ src/centralserver/admin-ui/src/store/modules/members.ts
@@ -1,21 +1,21 @@
-/**
+/*
  * The MIT License
- * <p>
+ *
  * Copyright (c) 2019- Nordic Institute for Interoperability Solutions (NIIS)
  * Copyright (c) 2018 Estonian Information System Authority (RIA),
  * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
  * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- * <p>
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * <p>
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * <p>
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -24,33 +24,20 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.niis.xroad.centralserver.restapi.service;
+import axios, {AxiosRequestConfig} from 'axios';
+import {
+  PagedClients,
+  PagingMetadata,
+  Client, ClientId,
+} from '@/openapi-types';
+import { defineStore } from 'pinia';
+import { DataOptions } from 'vuetify';
 
-import lombok.RequiredArgsConstructor;
-import org.niis.xroad.centralserver.restapi.entity.FlattenedSecurityServerClientView;
-import org.niis.xroad.centralserver.restapi.repository.FlattenedSecurityServerClientRepository;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class ClientService {
-
-    private final FlattenedSecurityServerClientRepository flattenedClientRepository;
-
-    private final StableSortHelper stableSortHelper;
-
-    public Page<FlattenedSecurityServerClientView> find(
-            FlattenedSecurityServerClientRepository.SearchParameters params,
-            Pageable pageable) {
-        Specification<FlattenedSecurityServerClientView> spec = flattenedClientRepository.multiParameterSearch(params);
-        pageable = stableSortHelper.addSecondaryIdSort(pageable);
-        return flattenedClientRepository.findAll(spec, pageable);
+export const memberStore = defineStore('member', {
+  actions: {
+    async add(client: Client) {
+      return axios
+        .post('/clients', client);
     }
-
-}
+  },
+});
